Move tests to better place
