+++ core/test/event_test.cpp
@@ -185,6 +185,46 @@ TEST_F(EventBroadcastTest, data_array_fail) {
   EXPECT_THROW(event::broadcast_weights(a), except::EventDataError);
 }
 
+static auto make_sparse() {
+  auto var = makeVariable<event_list<double>>(
+      Dims{Dim::Z, Dim::Y}, units::Unit(units::us), Shape{3, 2});
+  scipp::index count = 0;
+  for (auto &v : var.values<event_list<double>>())
+    v.resize(count++);
+  return var;
+}
+
+static auto make_sparse_with_variances() {
+  auto var = makeVariable<event_list<double>>(
+      Dimensions{{Dim::Z, 3}, {Dim::Y, 2}}, Values{}, Variances{});
+  scipp::index count = 0;
+  for (auto &v : var.values<event_list<double>>())
+    v.resize(count++);
+  count = 0;
+  for (auto &v : var.variances<event_list<double>>())
+    v.resize(count++);
+  return var;
+}
+
+TEST(EventSizesTest, fail_dense) {
+  auto bad = makeVariable<double>(Values{1.0});
+  EXPECT_ANY_THROW(event::sizes(bad));
+}
+
+TEST(EventSizesTest, no_variances) {
+  const auto var = make_sparse();
+  auto expected = makeVariable<scipp::index>(Dims{Dim::Z, Dim::Y}, Shape{3, 2},
+                                             Values{0, 1, 2, 3, 4, 5});
+  EXPECT_EQ(event::sizes(var), expected);
+}
+
+TEST(EventSizesTest, variances) {
+  const auto var = make_sparse_with_variances();
+  auto expected = makeVariable<scipp::index>(Dims{Dim::Z, Dim::Y}, Shape{3, 2},
+                                             Values{0, 1, 2, 3, 4, 5});
+  EXPECT_EQ(event::sizes(var), expected);
+}
+
 struct EventFilterTest : public ::testing::Test {
   Variable data_with_variances = makeVariable<event_list<float>>(
       Dims{Dim::X}, Shape{2}, units::Unit(units::counts),
